Simplification refactorings + primitive property deduplication
* Move property simplification from ProjectPropertiesReader to transformation
* Create PropertyDeduplicationTransformation
* Move code from ProjectWriter to DefaultAssemblyReferenceRemovalTransformation (closes #116)
* Fix for CPS projects `$(Configurations)` and `$(Platforms)` reading
* Fix and move around tests to match changes
* Move Pre and Post transformations to ConversionOptions
* Simplify ProjectConverter methods
