+++ Project2015To2017/ConversionOptions.cs
@@ -1,4 +1,6 @@
 using System.Collections.Generic;
+using System.Collections.Immutable;
+using Project2015To2017.Transforms;
 
 namespace Project2015To2017
 {
@@ -18,5 +20,13 @@ namespace Project2015To2017
 		/// Append the target framework to the output path
 		/// </summary>
 		public bool AppendTargetFrameworkToOutputPath { get; set; } = true;
+	    /// <summary>
+	    /// A collection of transforms executed before the execution of default ones
+	    /// </summary>
+	    public IReadOnlyCollection<ITransformation> PreDefaultTransforms { get; set; } = ImmutableArray<ITransformation>.Empty;
+	    /// <summary>
+	    /// A collection of transforms executed after the execution of default ones
+	    /// </summary>
+	    public IReadOnlyCollection<ITransformation> PostDefaultTransforms { get; set; } = ImmutableArray<ITransformation>.Empty;
 	}
 }
