+++ Project2015To2017Tests/ProjectPropertiesReadTest.cs
@@ -175,96 +175,6 @@ namespace Project2015To2017Tests
 		}
 
 		[TestMethod]
-		public async Task ReadsOptimize()
-		{
-			var xml = @"<?xml version=""1.0"" encoding=""utf-8""?>
-<Project ToolsVersion=""14.0"" DefaultTargets=""Build"" xmlns=""http://schemas.microsoft.com/developer/msbuild/2003"">
-  <PropertyGroup>
-    <OutputType>Library</OutputType>
-    <TargetFrameworkVersion>v4.6.2</TargetFrameworkVersion>
-    <Optimize>true</Optimize>
-  </PropertyGroup>
-</Project>";
-
-			var project = await ParseAndTransform(xml).ConfigureAwait(false);
-
-			Assert.AreEqual(true, project.Optimize);
-		}
-
-		[TestMethod]
-		public async Task ReadsTreatWarningsAsErrors()
-		{
-			var xml = @"<?xml version=""1.0"" encoding=""utf-8""?>
-<Project ToolsVersion=""14.0"" DefaultTargets=""Build"" xmlns=""http://schemas.microsoft.com/developer/msbuild/2003"">
-  <PropertyGroup>
-    <OutputType>Library</OutputType>
-    <TargetFrameworkVersion>v4.6.2</TargetFrameworkVersion>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-  </PropertyGroup>
-</Project>";
-
-			var project = await ParseAndTransform(xml).ConfigureAwait(false);
-
-			Assert.AreEqual(true, project.TreatWarningsAsErrors);
-		}
-
-		[TestMethod]
-		public async Task ReadsAllowUnsafeBlocks()
-		{
-			var xml = @"<?xml version=""1.0"" encoding=""utf-8""?>
-<Project ToolsVersion=""14.0"" DefaultTargets=""Build"" xmlns=""http://schemas.microsoft.com/developer/msbuild/2003"">
-  <PropertyGroup>
-    <OutputType>Library</OutputType>
-    <TargetFrameworkVersion>v4.6.2</TargetFrameworkVersion>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-</Project>";
-
-			var project = await ParseAndTransform(xml).ConfigureAwait(false);
-
-			Assert.AreEqual(true, project.AllowUnsafeBlocks);
-		}
-
-		[TestMethod]
-		public async Task ReadsSigningProperties()
-		{
-			var xml = @"<?xml version=""1.0"" encoding=""utf-8""?>
-<Project ToolsVersion=""14.0"" DefaultTargets=""Build"" xmlns=""http://schemas.microsoft.com/developer/msbuild/2003"">
-  <PropertyGroup>
-    <OutputType>Library</OutputType>
-    <TargetFrameworkVersion>v4.6.2</TargetFrameworkVersion>
-    <SignAssembly>true</SignAssembly>
-    <AssemblyOriginatorKeyFile>Project.snk</AssemblyOriginatorKeyFile>
-    <DelaySign>true</DelaySign>
-  </PropertyGroup>
-</Project>";
-
-			var project = await ParseAndTransform(xml).ConfigureAwait(false);
-
-			Assert.AreEqual(true, project.SignAssembly);
-			Assert.AreEqual("Project.snk", project.AssemblyOriginatorKeyFile);
-			Assert.AreEqual(true, project.DelaySign);
-		}
-
-		[TestMethod]
-		public async Task ReadsDefaultSigningProperties()
-		{
-			var xml = @"<?xml version=""1.0"" encoding=""utf-8""?>
-<Project ToolsVersion=""14.0"" DefaultTargets=""Build"" xmlns=""http://schemas.microsoft.com/developer/msbuild/2003"">
-  <PropertyGroup>
-    <OutputType>Library</OutputType>
-    <TargetFrameworkVersion>v4.6.2</TargetFrameworkVersion>
-  </PropertyGroup>
-</Project>";
-
-			var project = await ParseAndTransform(xml).ConfigureAwait(false);
-
-			Assert.AreEqual(false, project.SignAssembly);
-			Assert.AreEqual(null, project.AssemblyOriginatorKeyFile);
-			Assert.AreEqual(null, project.DelaySign);
-		}
-
-		[TestMethod]
 		[ExpectedException(typeof(NotSupportedException))]
 		public async Task ThrowsOnNoUnconditionalPropertyGroup()
 		{
@@ -642,142 +552,16 @@ if $(ConfigurationName) == Debug (
 			Assert.IsNull(project.AdditionalPropertyGroups[3].Attribute("Condition"));
 
 			var childrenDebug = project.AdditionalPropertyGroups[0].Elements().ToImmutableArray();
-			Assert.AreEqual(0, childrenDebug.Length);
-
-			var childrenRelease = project.AdditionalPropertyGroups[1].Elements().ToImmutableArray();
-			Assert.AreEqual(0, childrenRelease.Length);
-
-			var childrenReleaseCI = project.AdditionalPropertyGroups[2].Elements().ToImmutableArray();
-			// We remove only one property set to global default (FileAlignment)
-			Assert.AreEqual(7, childrenReleaseCI.Length);
-			Assert.IsTrue(ValidateChildren(childrenReleaseCI,
-				"DefineConstants", "OutputPath", "Optimize", "CodeAnalysisRuleSet", "DocumentationFile",
-				"TreatWarningsAsErrors", "RunCodeAnalysis"));
-			// check we are keeping original slashes and replacing configuration name with $(Configuration)
-			Assert.AreEqual(@"bin/$(Configuration)\", childrenReleaseCI.First(x => x.Name.LocalName == "OutputPath").Value);
-
-			var childrenGlobal = project.AdditionalPropertyGroups[3].Elements().ToImmutableArray();
-			Assert.AreEqual(0, childrenGlobal.Length);
-		}
-
-		[TestMethod]
-		public async Task SimplifiesProperties()
-		{
-			var xml = @"
-<Project ToolsVersion=""14.0"" DefaultTargets=""Build"" xmlns=""http://schemas.microsoft.com/developer/msbuild/2003"">
-  <PropertyGroup>
-    <Configuration Condition="" '$(Configuration)' == '' "">Debug</Configuration>
-    <Platform Condition="" '$(Platform)' == '' "">AnyCPU</Platform>
-    <ProjectGuid>{5C9DE16E-C69A-4182-9C0C-30FF7CC944CD}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Dopamine.Tests</RootNamespace>
-    <AssemblyName>Dopamine.Tests</AssemblyName>
-    <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
-    <FileAlignment>512</FileAlignment>
-    <ProjectTypeGuids>{3AC096D0-A1C2-E12C-1390-A8335801FDAB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
-    <VisualStudioVersion Condition=""'$(VisualStudioVersion)' == ''"">10.0</VisualStudioVersion>
-    <VSToolsPath Condition=""'$(VSToolsPath)' == ''"">$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)</VSToolsPath>
-    <ReferencePath>$(ProgramFiles)\Common Files\microsoft shared\VSTT\$(VisualStudioVersion)\UITestExtensionPackages</ReferencePath>
-    <IsCodedUITest>False</IsCodedUITest>
-    <TestProjectType>UnitTest</TestProjectType>
-    <SccProjectName>
-    </SccProjectName>
-    <SccLocalPath>
-    </SccLocalPath>
-    <SccAuxPath>
-    </SccAuxPath>
-    <SccProvider>
-    </SccProvider>
-    <NuGetPackageImportStamp>
-    </NuGetPackageImportStamp>
-  </PropertyGroup>
-  <PropertyGroup Condition="" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' "">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <PropertyGroup Condition="" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' "">
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-</Project>";
-
-			var project = await ParseAndTransform(xml).ConfigureAwait(false);
-
-			Assert.AreEqual(3, project.AdditionalPropertyGroups.Count);
-
-			Assert.IsNotNull(project.AdditionalPropertyGroups[0].Attribute("Condition"));
-			Assert.IsNotNull(project.AdditionalPropertyGroups[1].Attribute("Condition"));
-			Assert.IsNull(project.AdditionalPropertyGroups[2].Attribute("Condition"));
-
-			var childrenDebug = project.AdditionalPropertyGroups[0].Elements().ToImmutableArray();
 			Assert.AreEqual(2, childrenDebug.Length);
-			Assert.IsTrue(ValidateChildren(childrenDebug, "DebugType", "OutputPath"));
-			var childrenRelease = project.AdditionalPropertyGroups[1].Elements().ToImmutableArray();
-			Assert.AreEqual(2, childrenRelease.Length);
-			Assert.IsTrue(ValidateChildren(childrenRelease, "DebugType", "OutputPath"));
-			var childrenGlobal = project.AdditionalPropertyGroups[2].Elements().ToImmutableArray();
-			Assert.AreEqual(7, childrenGlobal.Length);
-			Assert.IsTrue(ValidateChildren(childrenGlobal,
-				"ProjectGuid", "ProjectTypeGuids", "VisualStudioVersion", "VSToolsPath",
-				"ReferencePath", "IsCodedUITest", "TestProjectType"));
-		}
 
-		[TestMethod]
-		public async Task HandlesComplexConditions()
-		{
-			var xml = @"
-<Project ToolsVersion=""14.0"" DefaultTargets=""Build"" xmlns=""http://schemas.microsoft.com/developer/msbuild/2003"">
-  <PropertyGroup>
-    <OutputType>Library</OutputType>
-    <AssemblyName>Dopamine.Tests</AssemblyName>
-    <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
-  </PropertyGroup>
-  <PropertyGroup Condition="" '$(Configuration)' == 'Debug' And '$(Platform)' == 'AnyCPU' "">
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-  </PropertyGroup>
-  <PropertyGroup Condition="" '$(Platform)|$(Configuration)' == 'AnyCPU|Release' "">
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-  </PropertyGroup>
-</Project>";
-
-			var project = await ParseAndTransform(xml).ConfigureAwait(false);
-
-			Assert.AreEqual(2, project.Configurations.Count);
-			Assert.AreEqual(1, project.Configurations.Count(x => x == "Debug"));
-			Assert.AreEqual(1, project.Configurations.Count(x => x == "Release"));
-
-			Assert.AreEqual(1, project.Platforms.Count);
-			Assert.AreEqual(1, project.Platforms.Count(x => x == "AnyCPU"));
-
-			Assert.AreEqual(3, project.AdditionalPropertyGroups.Count);
-
-			Assert.IsNotNull(project.AdditionalPropertyGroups[0].Attribute("Condition"));
-			Assert.IsNotNull(project.AdditionalPropertyGroups[1].Attribute("Condition"));
-			Assert.IsNull(project.AdditionalPropertyGroups[2].Attribute("Condition"));
-
-			var childrenDebug = project.AdditionalPropertyGroups[0].Elements().ToImmutableArray();
-			Assert.AreEqual(2, childrenDebug.Length);
-			Assert.IsTrue(ValidateChildren(childrenDebug, "DebugType", "OutputPath"));
 			var childrenRelease = project.AdditionalPropertyGroups[1].Elements().ToImmutableArray();
 			Assert.AreEqual(1, childrenRelease.Length);
-			Assert.IsTrue(ValidateChildren(childrenRelease, "OutputPath"));
-			var childrenGlobal = project.AdditionalPropertyGroups[2].Elements().ToImmutableArray();
-			Assert.AreEqual(0, childrenGlobal.Length);
+
+			var childrenReleaseCI = project.AdditionalPropertyGroups[2].Elements().ToImmutableArray();
+			Assert.AreEqual(8, childrenReleaseCI.Length);
+
+			var childrenGlobal = project.AdditionalPropertyGroups[3].Elements().ToImmutableArray();
+			Assert.AreEqual(4, childrenGlobal.Length);
 		}
 
 		private static async Task<Project> ParseAndTransform(
